requirements.txt
streamlit>=1.28.0
fr24sdk>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
